Reorganize files for better clarity
