+++ app/src/main/java/com/nqueens/game/MainActivity.kt
@@ -4,7 +4,6 @@ import android.os.Bundle
 import androidx.activity.ComponentActivity
 import androidx.activity.compose.setContent
 import androidx.activity.enableEdgeToEdge
-import com.nqueens.game.ui.ChessGamesApp
 import dagger.hilt.android.AndroidEntryPoint
 
 @AndroidEntryPoint
