+++ app/src/main/java/com/nqueens/game/core/design/theme/Type.kt
@@ -1,4 +1,4 @@
-package com.nqueens.game.ui.theme
+package com.nqueens.game.core.design.theme
 
 import androidx.compose.material3.Typography
